#26744  fixing condition
