+++ tools/dotcms-cli/cli/src/test/java/com/dotcms/cli/command/PushCommandIT.java
@@ -301,40 +301,4 @@ class PushCommandIT extends CommandTest {
         }
     }
 
-    /**
-     * This helper method is used to create a temporary folder for the test.
-     *
-     * @return a {@link Path} object representing a temporary directory for the test.
-     * @throws IOException if there's a problem in creating the temporary directory.
-     */
-    private Path createTempFolder() throws IOException {
-
-        String randomFolderName = "folder-" + UUID.randomUUID();
-        return Files.createTempDirectory(randomFolderName);
-    }
-
-    /**
-     * This helper method is used to delete a directory and all its contents.
-     *
-     * @param folderPath the {@link Path} object of the directory to delete.
-     * @throws IOException if there's a problem in deleting the directory or its contents.
-     */
-    private void deleteTempDirectory(Path folderPath) throws IOException {
-        Files.walkFileTree(folderPath, new SimpleFileVisitor<Path>() {
-            @Override
-            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs)
-                    throws IOException {
-                Files.delete(file); // Deletes the file
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult postVisitDirectory(Path dir, IOException exc)
-                    throws IOException {
-                Files.delete(dir); // Deletes the directory after its content has been deleted
-                return FileVisitResult.CONTINUE;
-            }
-        });
-    }
-
 }
