+++ tools/dotcms-cli/cli/src/test/java/com/dotcms/cli/command/contenttype/ContentTypeCommandIT.java
@@ -27,15 +27,11 @@ import io.quarkus.test.junit.TestProfile;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
-import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.util.Date;
 import java.util.List;
 import java.util.Objects;
-import java.util.UUID;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
@@ -436,7 +432,9 @@ class ContentTypeCommandIT extends CommandTest {
      */
     @Test
     void Test_Pull_Same_Content_Type_Multiple_Times() throws IOException {
-        final Workspace workspace = workspaceManager.getOrCreate();
+        // Create a temporal folder for the workspace
+        var tempFolder = createTempFolder();
+        final Workspace workspace = workspaceManager.getOrCreate(tempFolder);
         final CommandLine commandLine = createCommand();
         final StringWriter writer = new StringWriter();
         try (PrintWriter out = new PrintWriter(writer)) {
@@ -908,42 +906,6 @@ class ContentTypeCommandIT extends CommandTest {
         } finally {
             deleteTempDirectory(tempFolder);
         }
-    }
-
-    /**
-     * Creates a temporary folder with a random name.
-     *
-     * @return The path to the created temporary folder.
-     * @throws IOException If an I/O error occurs while creating the temporary folder.
-     */
-    private Path createTempFolder() throws IOException {
-
-        String randomFolderName = "folder-" + UUID.randomUUID();
-        return Files.createTempDirectory(randomFolderName);
-    }
-
-    /**
-     * Deletes a temporary directory and all its contents.
-     *
-     * @param folderPath The path to the temporary directory to be deleted.
-     * @throws IOException If an I/O error occurs while deleting the directory or its contents.
-     */
-    private void deleteTempDirectory(Path folderPath) throws IOException {
-        Files.walkFileTree(folderPath, new SimpleFileVisitor<>() {
-            @Override
-            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs)
-                    throws IOException {
-                Files.delete(file); // Deletes the file
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult postVisitDirectory(Path dir, IOException exc)
-                    throws IOException {
-                Files.delete(dir); // Deletes the directory after its content has been deleted
-                return FileVisitResult.CONTINUE;
-            }
-        });
     }
 
 }
